# openapi/mcp-oxide.openapi.yaml
openapi: 3.1.0
info:
  title: mcp-oxide Gateway API
  description: |
    Control plane + data plane for mcp-oxide. In Phase 0 only the health
    endpoints are stable; subsequent phases introduce `/adapters`, `/tools`,
    `/mcp`, and `/adapters/{name}/mcp`.
  version: 0.1.0
  license:
    name: Apache-2.0 OR MIT
servers:
  - url: http://localhost:8080
    description: Local
tags:
  - name: health
  - name: control-plane
  - name: data-plane

paths:
  /healthz:
    get:
      tags: [health]
      summary: Liveness + provider summary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }

  /readyz:
    get:
      tags: [health]
      summary: Readiness
      responses:
        "200": { description: Ready }
        "503": { description: Not ready }

  /livez:
    get:
      tags: [health]
      summary: Liveness
      responses:
        "200": { description: Live }

  # -------- Control plane (Phase 2+) ----------------------------------------
  /adapters:
    get:
      tags: [control-plane]
      summary: List adapters
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Adapter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [control-plane]
      summary: Create + deploy adapter
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Adapter" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adapter" }
        "400": { $ref: "#/components/responses/InvalidParams" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }

  /adapters/{name}:
    parameters:
      - $ref: "#/components/parameters/Name"
    get:
      tags: [control-plane]
      summary: Read adapter
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adapter" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [control-plane]
      summary: Update adapter
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Adapter" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adapter" }
    delete:
      tags: [control-plane]
      summary: Delete adapter
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  /adapters/{name}/status:
    parameters:
      - $ref: "#/components/parameters/Name"
    get:
      tags: [control-plane]
      summary: Deployment status
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeploymentStatus" }

  /adapters/{name}/logs:
    parameters:
      - $ref: "#/components/parameters/Name"
    get:
      tags: [control-plane]
      summary: Stream deployment logs (SSE)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: SSE stream of log lines
          content:
            text/event-stream: {}

  /tools:
    get:
      tags: [control-plane]
      summary: List tools
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tool" }
    post:
      tags: [control-plane]
      summary: Register + deploy tool
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Tool" }
      responses:
        "201": { description: Created }

  /tools/{name}:
    parameters:
      - $ref: "#/components/parameters/Name"
    get:
      tags: [control-plane]
      summary: Read tool
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tool" }
    put:
      tags: [control-plane]
      summary: Update tool
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Tool" }
      responses:
        "200": { description: Updated }
    delete:
      tags: [control-plane]
      summary: Delete tool
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }

  # -------- Data plane (Phase 1+) -------------------------------------------
  /mcp:
    post:
      tags: [data-plane]
      summary: MCP JSON-RPC via Tool Gateway Router
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonRpcRequest" }
      responses:
        "200":
          description: MCP JSON-RPC response (or SSE stream)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JsonRpcResponse" }
            text/event-stream: {}

  /adapters/{name}/mcp:
    parameters:
      - $ref: "#/components/parameters/Name"
    post:
      tags: [data-plane]
      summary: MCP JSON-RPC for a specific adapter
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonRpcRequest" }
      responses:
        "200":
          description: MCP response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JsonRpcResponse" }
            text/event-stream: {}

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Name:
      in: path
      name: name
      required: true
      schema: { type: string, pattern: "^[a-z0-9][a-z0-9-]{0,62}$" }

  responses:
    Unauthenticated:
      description: Missing/invalid token
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Policy deny
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    InvalidParams:
      description: Invalid request
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }

  schemas:
    Health:
      type: object
      required: [status, version]
      properties:
        status: { type: string, enum: [ok] }
        version: { type: string }
        providers: { type: object, additionalProperties: { type: string } }

    Error:
      type: object
      required: [error, message]
      properties:
        error:   { type: string }
        message: { type: string }

    Adapter:
      type: object
      required: [name, image]
      properties:
        name:             { type: string }
        description:      { type: string }
        image:            { type: string, description: "OCI image reference" }
        endpoint_port:    { type: integer, default: 8080 }
        endpoint_path:    { type: string, default: "/mcp" }
        replicas:         { type: integer, minimum: 0, default: 1 }
        env:
          type: array
          items: { $ref: "#/components/schemas/EnvVar" }
        secret_refs:
          type: array
          items: { $ref: "#/components/schemas/SecretRef" }
        required_roles:   { type: array, items: { type: string } }
        tags:             { type: array, items: { type: string } }
        resources:        { $ref: "#/components/schemas/Resources" }
        health:           { $ref: "#/components/schemas/HealthProbe" }
        session_affinity: { type: string, enum: [sticky, none], default: sticky }
        labels:           { type: object, additionalProperties: { type: string } }
        revision:         { type: integer, description: "Optimistic concurrency version" }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }

    Tool:
      type: object
      required: [name, image, tool_definition]
      properties:
        name:             { type: string }
        description:      { type: string }
        image:            { type: string, description: "OCI image reference" }
        endpoint_port:    { type: integer, default: 8080 }
        endpoint_path:    { type: string, default: "/mcp" }
        tool_definition:  { $ref: "#/components/schemas/ToolDefinition" }
        env:
          type: array
          items: { $ref: "#/components/schemas/EnvVar" }
        secret_refs:
          type: array
          items: { $ref: "#/components/schemas/SecretRef" }
        required_roles:   { type: array, items: { type: string } }
        tags:             { type: array, items: { type: string } }
        resources:        { $ref: "#/components/schemas/Resources" }
        revision:         { type: integer, description: "Optimistic concurrency version" }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }

    ImageRef:
      type: object
      required: [reference]
      properties:
        reference: { type: string, example: "registry.example.com/mcp-aws:1.4.0" }

    Endpoint:
      type: object
      required: [port]
      properties:
        port: { type: integer }
        path: { type: string, default: "/mcp" }

    EnvVar:
      type: object
      required: [name, value]
      properties:
        name:  { type: string }
        value: { type: string }

    SecretRef:
      type: object
      required: [name, provider, key]
      properties:
        name:     { type: string }
        provider: { type: string }
        key:      { type: string }

    Resources:
      type: object
      properties:
        cpu:    { type: string, example: "500m" }
        memory: { type: string, example: "512Mi" }

    HealthProbe:
      type: object
      required: [path, port]
      properties:
        path: { type: string }
        port: { type: integer }

    ToolDefinition:
      type: object
      required: [name, input_schema]
      properties:
        name:         { type: string }
        title:        { type: string }
        description:  { type: string }
        input_schema: { type: object }
        annotations:  { type: object }

    DeploymentStatus:
      type: object
      properties:
        ready:           { type: boolean }
        replicas:        { type: integer }
        ready_replicas:  { type: integer }
        message:         { type: string }

    JsonRpcRequest:
      type: object
      required: [jsonrpc, method]
      properties:
        jsonrpc: { type: string, enum: ["2.0"] }
        id:      {}
        method:  { type: string }
        params:  {}

    JsonRpcResponse:
      type: object
      required: [jsonrpc]
      properties:
        jsonrpc: { type: string, enum: ["2.0"] }
        id:      {}
        result:  {}
        error:   { $ref: "#/components/schemas/JsonRpcError" }

    JsonRpcError:
      type: object
      required: [code, message]
      properties:
        code:    { type: integer }
        message: { type: string }
        data:    {}